Review comment
- Move multiplyQuantities and divideQuantities to Math.kt
- Have them as operators
